+++ prisma/sql/setWorkerMemory.sql
@@ -1,1 +1,0 @@
-set maintenance_work_mem='10 GB';